revision based on comments and refactoring to add utils folder
