+++ generator/checker_test.go
@@ -15,15 +15,15 @@
 package generator
 
 import (
-	openapiv3 "github.com/googleapis/gnostic/openapiv3"
-	plugins "github.com/googleapis/gnostic/plugins"
-	"os/exec"
 	"testing"
+
+	"github.com/googleapis/gnostic-grpc/utils"
+	plugins "github.com/googleapis/gnostic/plugins"
 )
 
 func TestNewFeatureCheckerParameters(t *testing.T) {
 	input := "testfiles/parameters.yaml"
-	documentv3, err := ParseOpenAPIDoc(input)
+	documentv3, err := utils.ParseOpenAPIDoc(input)
 	if err != nil {
 		t.Errorf("Error while parsing input file: %s", input)
 		return
@@ -43,7 +43,7 @@ func TestNewFeatureCheckerParameters(t *testing.T) {
 
 func TestFeatureCheckerRequestBodies(t *testing.T) {
 	input := "testfiles/requestBodies.yaml"
-	documentv3, err := ParseOpenAPIDoc(input)
+	documentv3, err := utils.ParseOpenAPIDoc(input)
 	if err != nil {
 		t.Errorf("Error while parsing input file: %s", input)
 		return
@@ -62,7 +62,7 @@ func TestFeatureCheckerRequestBodies(t *testing.T) {
 
 func TestFeatureCheckerResponses(t *testing.T) {
 	input := "testfiles/responses.yaml"
-	documentv3, err := ParseOpenAPIDoc(input)
+	documentv3, err := utils.ParseOpenAPIDoc(input)
 	if err != nil {
 		t.Errorf("Error while parsing input file: %s", input)
 		return
@@ -81,7 +81,7 @@ func TestFeatureCheckerResponses(t *testing.T) {
 
 func TestFeatureCheckerOther(t *testing.T) {
 	input := "testfiles/other.yaml"
-	documentv3, err := ParseOpenAPIDoc(input)
+	documentv3, err := utils.ParseOpenAPIDoc(input)
 	if err != nil {
 		t.Errorf("Error while parsing input file: %s", input)
 		return
@@ -110,14 +110,4 @@ func validateKeys(t *testing.T, expectedKeys [][]string, messages []*plugins.Mes
 			}
 		}
 	}
-}
-
-func ParseOpenAPIDoc(input string) (*openapiv3.Document, error) {
-	cmd := exec.Command("gnostic", "--pb-out=-", input)
-	b, err := cmd.Output()
-	if err != nil {
-		return nil, err
-	}
-	documentv3, err := createOpenAPIDocFromGnosticOutput(b)
-	return documentv3, err
 }
