+++ incompatibility/incompatibility-report/incomp-scanning.go
@@ -18,11 +18,11 @@ import (
 )
 
 // Scan for incompatibilities in an OpenAPI document
-func getIncompatibilites(document *openapiv3.Document) *IncompatibilityReport {
+func ScanIncompatibilities(document *openapiv3.Document) *IncompatibilityReport {
 	var incompatibilities []*Incompatibility
 
-	if document.Servers != nil { // Simple servers lookup
-		incompatibilities = append(incompatibilities, &Incompatibility{Token: "servers", Classification: "SERVERS"})
+	if document.Servers != nil {
+		incompatibilities = append(incompatibilities, &Incompatibility{TokenPath: []string{"servers"}, Classification: "SERVERS"})
 	}
 	return &IncompatibilityReport{Incompatibilities: incompatibilities}
 }
